Update int -> double

+++ dto/lab.go
@@ -4,7 +4,7 @@ import "time"
 
 type CreateBreathingHistoryRequest struct {
 	IdForLogin    string    `json:"id_for_login"`
-	AverageVolume int       `json:"average_volume"`
+	AverageVolume float64   `json:"average_volume"`
 	Timestamp     time.Time `json:"timestamp"`
 }
